% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{estimateR}
\alias{estimateR}
\title{Parameter estimates of 2PL IRT model by MCMC method based on Gibbs sampler.}
\usage{
estimateR(x, n_chain = 3, burn = 1000, n_itr = 3000)
}
\arguments{
\item{x}{dataset include dichotomous responses to items of subjects}

\item{n_chain}{number of chain}

\item{burn}{burnin length}

\item{n_itr}{length of each chain}
}
\value{
estimates of theta, a and b
}
\description{
A Gibbs sampler using R. Due to existence of burn-in period, we take part of sample values and take the average. What's more, the following function permits several chains to ensure convergence.
}
\examples{
\dontrun{
library(ltm)
est<-estimateR(t(LSAT))
est$theta
est$a
est$b
}
}
