% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_mcmc.R
\name{gibbs_mcmc}
\alias{gibbs_mcmc}
\title{Posterior estimate}
\usage{
gibbs_mcmc(x, n_itr)
}
\arguments{
\item{x}{dataset include dichotomous responses to items of subjects}

\item{n_itr}{length of each chain}
}
\value{
posterior estimates of theta,a and b for n_itr iterations.
}
\description{
Function to obtain posterior estimates of theta,a and b.
}
\examples{
\dontrun{
librar(ltm)
gibbs_mcmc(t(LSAT),3e3)
}
}
