% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_2pl.R
\name{irt_2pl}
\alias{irt_2pl}
\title{compute item response probability}
\usage{
irt_2pl(a, b, theta)
}
\arguments{
\item{a}{discrimination parameter of an item}

\item{b}{difficulty parameters of an item}

\item{theta}{ability parameter of a participant/subject/student}
}
\description{
Under 2PL IRT model, we consider responses to item are distichous, thus, the probability that one subject with ability theta response correctly to an item with dicrimination 'a' and difficulty 'b' is P(X=1)=expit(a(theta-b))
}
\examples{
\dontrun{
irt_2pl(1,0.5,2)
}
}
