% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglikC}
\alias{loglikC}
\title{Compute the log-likelihood function  using Rcpp}
\usage{
loglikC(theta, a, b, x)
}
\arguments{
\item{theta}{ability parameter of a participant/subject/student}

\item{a}{discriminationparameter of an item}

\item{b}{difficulty parameters of an item}

\item{x}{subject's response to an item}
}
\description{
Compute the log-likelihood function  using Rcpp
}
\examples{
\dontrun{
lp<-loglikC(2,0.5,1,1)
}
}
