% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_a.R
\name{update_a}
\alias{update_a}
\title{Update a}
\usage{
update_a(a, theta, b, x, sigma)
}
\arguments{
\item{a}{discriminationparameter of an item}

\item{theta}{ability parameter of a participant/subject/student}

\item{b}{difficulty parameters of an item}

\item{x}{subject's response to an item}

\item{sigma}{standard error of random normal distribution that is set in advance}
}
\description{
Function written for updating discrimination parameter 'a'.
}
