% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_b.R
\name{update_b}
\alias{update_b}
\title{Update b}
\usage{
update_b(b, theta, a, x, sigma)
}
\arguments{
\item{b}{difficulty parameters of an item}

\item{theta}{ability parameter of a participant/subject/student}

\item{a}{discriminationparameter of an item}

\item{x}{subject's response to an item}

\item{sigma}{standard error of random normal distribution that is set in advance}
}
\description{
Function written for updating difficulty parameter 'b'.
}
