% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_theta.R
\name{update_theta}
\alias{update_theta}
\title{Update theta}
\usage{
update_theta(theta, a, b, x, sigma)
}
\arguments{
\item{theta}{ability parameter of a participant/subject/student}

\item{a}{discriminationparameter of an item}

\item{b}{difficulty parameters of an item}

\item{x}{subject's response to an item}

\item{sigma}{standard error of random normal distribution that is set in advance}
}
\description{
We use gibbs sampler to obtain estimates of parameters, refer to the book Item Response Theory Parameter Estimation Techniques, 2rd edition. During samples generation, we need to update theta,a,b respectively. The following three functions are displayed.Function written for updating ability parameter 'theta'.
}
